+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## Unreleased
+## 1.3.0
 * Fix: Ensure REST response on SQL Import.
 * Feat: Add Progress bar to Parse activity.
 * Feat: Implement Purge component.
-* Refactor: Move Interval logic to Progress Bar component.
+* Refactor: Move Interval logic to ProgressBar component.
+* Refactor: Move Handle logic away from App to ImportButton component.
+* Chore: Update doc blocks for components.
+* Tested up to WP 6.7.1
 
 ## 1.2.2
 * Refactor: Parser instance via DI logic.